requirements.txt
asgiref==3.3.4
certifi==2020.12.5
Django==3.2.3
djangorestframework==3.12.4
gunicorn==20.1.0
psycopg2-binary==2.8.6
pytz==2021.1
python-json-logger==2.0.1
sqlparse==0.4.1
kubernetes
django-extensions
pydotplus
whitenoise
django-allow-cidr
